Apply some review fixes
