+++ lib/ide/src/entry_point.rs
@@ -1,6 +1,6 @@
 #![allow(missing_docs)]
 
-use super::view::project_view::ProjectView;
+use super::view::project::ProjectView;
 
 pub fn entry_point() {
     ProjectView::new().forget();
