+++ lib/ide/src/view.rs
@@ -10,6 +10,6 @@
 #![warn(unused_qualifications)]
 
 pub mod ui_component;
-pub mod project_view;
-pub mod view_layout;
+pub mod project;
+pub mod layout;
 pub mod text_editor;
