+++ lib/ide/src/view/layout.rs
